begin custom view for call to action

+++ app/src/main/java/com/zuk0/gaijinsmash/riderz/data/local/entity/weather_response/Coordinates.kt
@@ -6,9 +6,9 @@ import com.google.gson.annotations.SerializedName
 data class Coordinates(
         @Expose
         @SerializedName("lon")
-        var longitude: Long,
+        var longitude: Double,
 
         @Expose
         @SerializedName("lat")
-        var latitude: Long
+        var latitude: Double
 )
